+++ history.md
@@ -2,10 +2,10 @@
 ## 2.0.0(*)
 * feat: 接入 `yyl-concat-webpack-plugin`, `yyl-copy-webpack-plugin`, `yyl-sugar-webpack-plugin`, `yyl-rev-webpack-plugin`
 * feat: 接入 `happypack`
+* feat: 整个 webpack 支持 ts
 * feat: 升级 webpack 到 `4.41.6`
 * feat: `node-sass` 废弃， 改用 `sass`
 * feat: `uglify-webpack-plugin` 废弃， 改用 `terser-webpack-plugin`
-* feat: 迁移 `@babel/polyfill` 到具体 pack
 * feat: test-case 覆盖 copy， concat， rev 用例
 * feat: 单元测试框架变更 `jest` -> `mocha`
 * del: 去掉 `typescript`, `react-ts-ie8` 配置项